sunone_aimbot_cpp/China_ai/requirements.txt
# ZTXAI Dependencies
# Generated from _internal folder analysis
# Install: pip install -r requirements.txt

# Computer Vision & Image Processing
opencv-python>=4.5.0
numpy>=1.19.0
Pillow>=8.0.0
bettercam>=0.3.0

# GUI Framework
dearpygui>=1.6.0

# Plotting (found in _internal: matplotlib, contourpy, kiwisolver)
matplotlib>=3.5.0

# Input Control
pynput>=1.7.6
pyclick==0.0.2
pydirectinput>=1.0.4

# Windows API (found in _internal: win32, Pythonwin, pywin32_system32)
pywin32>=300
WMI>=1.5.1

# Network & HTTP (found in _internal: certifi, charset_normalizer)
requests>=2.25.0
websocket-client>=0.57.0
certifi>=2021.0.0

# Encryption (found in _internal: Crypto, cryptography, cffi)
cryptography>=3.4.0
pycryptodome>=3.14.0
arc4>=0.1.0
cffi>=1.14.0

# Machine Learning / Inference
# NOTE: Choose ONE of these (cannot install both):
#   - onnxruntime        = CPU only
#   - onnxruntime-gpu    = GPU (CUDA 12.x) + CPU fallback
# For CUDA 11.x use: pip install onnxruntime-gpu --index-url https://aiinfra.pkgs.visualstudio.com/PublicPackages/_packaging/onnxruntime-cuda-11/pypi/simple/
#onnxruntime-gpu>=1.8.0
scipy>=1.5.0
filterpy>=1.4.2

# Serial Communication
pyserial>=3.5

# XML Processing (found in _internal: lxml)
lxml>=4.6.0

# YAML Processing (found in _internal: yaml)
pyyaml>=5.4.0

# Compression (found in _internal: zstandard)
zstandard>=0.15.0

# Interfaces (found in _internal: zope)
zope.interface>=5.0.0

# Build Tools
pyinstaller>=5.0.0
cython>=0.29.0
